added test files, and folders, added additional error handling

+++ migrations/sqls/20230516160157-storefront-tables-up.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS users_table (
 CREATE TABLE IF NOT EXISTS products_table (
     id SERIAL PRIMARY KEY NOT NULL,
     product_name VARCHAR(100) NOT NULL,
-    price INTEGER NOT NULL,
+    price FLOAT NOT NULL,
     product_category VARCHAR(100) NOT NULL
 );
